requirements.txt
absl-py==0.12.0
ale-py==0.7.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.1.0
astunparse==1.6.3
attrs==22.2.0
autopep8==1.6.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chainmap==1.0.3
chardet==4.0.0
click==8.0.3
cloudpickle==2.0.0
combomethod==1.0.12
configparser==5.0.2
contourpy==1.0.5
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.24
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.12
dropbox==11.20.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.0
flake8==6.1.0
flatbuffers==1.12
fonttools==4.37.4
fqdn==1.5.1
gast==0.4.0
gensim==4.2.0
gitdb==4.0.7
GitPython==3.1.7
glfw==2.4.0
gnureadline==8.0.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.53.0
gym==0.21.0
h5py==3.1.0
huggingface-hub==0.10.0
idna==2.10
imageio==2.10.3
importlib-metadata==4.6.1
importlib-resources==5.4.0
iniconfig==1.1.1
IProgress==0.4
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.1
jeepney==0.6.0
Jinja2==3.1.2
jira==3.0.1
joblib==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.4
jupyter-console==6.6.2
jupyter_core==5.2.0
jupyter-events==0.6.3
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
lightgbm==3.3.2
lizard==1.17.7
lockfile==0.12.2
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.1.2
matplotlib==3.6.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mypy==1.1.1
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
nulltype==2.3.1
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==23.0
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==22.3.1
platformdirs==3.0.0
pluggy==1.0.0
ply==3.11
progressbar2==3.53.1
prometheus-client==0.16.0
prompt-toolkit==3.0.32
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.0
pycparser==2.20
PyDriller==1.15.2
pyflakes==3.1.0
PyGithub==1.55
pyglet==1.5.21
Pygments==2.13.0
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.19.3
pytest==7.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-utils==2.5.6
pytz==2021.1
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.9.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
scikit-learn==1.1.2
scipy==1.6.3
seaborn==0.11.1
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.99
setuptools==65.6.3
simplejson==3.17.2
six==1.16.0
sklearn==0.0
smart-open==6.2.0
smmap==4.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.19
stack-data==0.6.0
stone==3.2.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.17.1
testresources==2.0.1
threadpoolctl==2.1.0
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.10.2
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.22.2
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.4
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==1.0.1
wheel==0.37.1
widgetsnbextension==4.0.5
wrapt==1.12.1
zipp==3.5.0
